+++ exchangelib/autodiscover/discovery.py
@@ -75,6 +75,9 @@ class Autodiscovery:
     INITIAL_RETRY_POLICY = FailFast()
     RETRY_WAIT = 10  # Seconds to wait before retry on connection errors
     MAX_REDIRECTS = 10  # Maximum number of URL redirects before we give up
+    DNS_RESOLVER_ATTRS = {
+        'timeout': AutodiscoverProtocol.TIMEOUT,
+    }
 
     def __init__(self, email, credentials=None, auth_type=None, retry_policy=None):
         """
@@ -149,7 +152,8 @@ class Autodiscovery:
     @threaded_cached_property
     def resolver(self):
         resolver = dns.resolver.Resolver()
-        resolver.timeout = AutodiscoverProtocol.TIMEOUT
+        for k, v in self.DNS_RESOLVER_ATTRS.items():
+            setattr(resolver, k, v)
         return resolver
 
     def _build_response(self, ad_response):